Used error value when weather sensor is unreadable
Temperature, pressure, and humidity values need to be initialized when
there is an error reading the weather sensor. This way, GPS readings may
still be recorded even when the sensor is non-functional or missing.

This commit adds reading_error value of 999.99c which will be used in the
event of the weather sensor data being unavailable.

Signed-off-by: Mike Szczys <mike@golioth.io>