Only stream valid sensor readings
Add checks for to compare sensor reading to the stored reading_error values
and do not stream those to Golioth if they match. If for any reason the
weather sensor cannot be read (missing or damaged) the sensor values will
be absent from the data packet set to the cloud.

Signed-off-by: Mike Szczys <mike@golioth.io>